CUDA 11.0 instead of 11.1 for TF 2.4 compatibility; also named model outputs for huggingface >= 3.0.0 (#47)
Change optimizer.loss_scale() to optimizer.loss_scale for TF 2.4. Fixes #44.

I ran pretraining on ALBERT for a few thousand steps with XLA enabled on TF 2.4 and observed the loss drop from 11 to 7.5 in 3k steps with global batch size 32. I am not able to replicate #45. It is possible that fixing the named model outputs has solved this issue.